20강, 수정?id=1

+++ src/main/java/com/ll/domain/WiseSaying/WiseSayingController.java
@@ -2,6 +2,7 @@ package com.ll.domain.WiseSaying;
 
 import com.ll.AppContext;
 import com.ll.WiseSaying;
+import com.ll.global.rq.Rq;
 
 import java.util.*;
 import java.util.stream.Collectors;
@@ -16,10 +17,10 @@ public class WiseSayingController {
     }
 
     public void actionWrite() {
-        System.out.println("명언 : ");
+        System.out.print("명언 : ");
         String content = scanner.nextLine();
 
-        System.out.println("작가 : ");
+        System.out.print("작가 : ");
         String author = scanner.nextLine();
 
         WiseSaying wiseSaying = wiseSayingService.write(content, author);
@@ -35,18 +36,13 @@ public class WiseSayingController {
         }
     }
 
-    public void actionDelete(String cmd) {
-        String[] cmdBits = cmd.split("\\?", 2);
-        String queryString = cmdBits[1];
+    public void actionDelete(Rq rq) {
+        int id = rq.getParamsAsInt("id", -1);
 
-        Map<String, String> params = Arrays
-                .stream(queryString.split("&"))
-                .map(e -> e.split("=", 2))
-                .filter(e -> e.length == 2 && !e[0].isBlank() && !e[1].isBlank())
-                .collect(Collectors.toMap(e -> e[0].trim(), e -> e[1].trim()));
-
-        String idStr = params.getOrDefault("id", "");
-        int id = Integer.parseInt(idStr);
+        if (id == -1 ) {
+            System.out.println("id를 숫자로 입력해주세요.");
+            return;
+        }
 
         boolean isDeleted = wiseSayingService.delete(id);
 
@@ -56,5 +52,31 @@ public class WiseSayingController {
         }
 
         System.out.printf("%d번 명언이 삭제되었습니다.\n", id);
+    }
+
+    public void actionModify(Rq rq) {
+        int id = rq.getParamsAsInt("id", -1);
+
+        if (id == -1 ) {
+            System.out.println("id를 숫자로 입력해주세요.");
+            return;
+        }
+
+        WiseSaying wiseSaying = wiseSayingService.findById(id);
+
+        if (wiseSaying == null) {
+            System.out.printf("%d번 명언은 존재하지 않습니다.\n", id);
+            return;
+        }
+
+        System.out.printf("명언 (기존) : %s\n", wiseSaying.getContent());
+        System.out.print("명언 : ");
+        String modifyContent = scanner.nextLine();
+
+        System.out.printf("작가 (기존) : %s\n", wiseSaying.getAuthor());
+        System.out.print("작가 : ");
+        String modifyAuthor = scanner.nextLine();
+
+        wiseSayingService.modify(wiseSaying, modifyContent, modifyAuthor);
     }
 }
